requirements.txt
streamlit
pandas
numpy
pillow
plotly
tensorflow
matplotlib
seaborn
scikit-learn
